third_party/blink/web_tests/external/wpt/css/css-align/blocks/align-content-block-008-expected.txt

This is a testharness.js-based test.
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 10
Harness: the test ran to completion.
